Assignment6/requirements.txt
mcp
Pillow
google-genai
sympy
numpy